'Revise file structure to accomidate back-end Spotify authorization'
