+++ README.md
@@ -1,3 +1,1 @@
 # YouTube Mixtape
-
-[View here](https://rawgit.com/VitaC123/youTubeMixTape/master/index.html)
