+++ package.json
@@ -1,0 +1,30 @@
+{
+  "name": "YouTube Mixtape",
+  "version": "0.2.0",
+  "description": "Combining Spotify's genre search with YouTube video results",
+  "author": "chris vita <chris.vita@gmail.com>",
+  "main": "server.js",
+  "scripts": {
+    "start": "nodemon server.js"
+  },
+  "dependencies": {
+    "express": "4.15.2",
+    "request": "^2.81.0"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/VitaC123/YouTube-Mixtape"
+  },
+  "keywords": [
+    "youtube",
+    "spotify",
+    "api",
+    "jquery",
+    "node",
+    "express"
+  ],
+  "license": "MIT",
+  "devDependencies": {
+    "nodemon": "^1.11.0"
+  }
+}
