+++ views/index.html
@@ -130,6 +130,7 @@
   <script src="https://apis.google.com/js/platform.js" async defer></script>
   <script src="https://apis.google.com/js/client.js?onload=googleApiClientReady"></script>
 
+  <script src="js/authSpotify.js"></script>
   <script src="js/searchSpotify.js"></script>
   <script src="js/searchYouTube.js"></script>
