+++ views/js/playlistsInsert.js
